Create Westerhout_2024_PFOA_PFOS_PBK_model.R
This model code is not the same as previously published, because that code contained a mistake in the calculated exposure during the first year of life. This has been corrected. In addition, the (free) concentration in plasma is calculated differently compared to the published model in order to have a correct mass balance.